+++ Eventy/app/src/main/java/com/example/eventy/home/events/featured_events/FeaturedEventsFragment.java
@@ -11,7 +11,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.eventy.adapters.FeaturedEventsAdapter;
+import com.example.eventy.adapters.events.FeaturedEventsAdapter;
 import com.example.eventy.databinding.FragmentHomeFeaturedEventsBinding;
 import com.example.eventy.model.enums.PrivacyType;
 import com.example.eventy.model.event.Event;
